# bradobrei_party/backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API для системы управления сетью барбершопов Bradobrei Party.",
        "title": "Bradobrei Party API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает все услуги сети.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Список услуг",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Service"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новую услугу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Создать услугу",
                "parameters": [
                    {
                        "description": "Данные услуги",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Service"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Service"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает услуги, привязанные к текущему мастеру.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Мои услуги",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Service"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает одну услугу по идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Услуга по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Service"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет существующую услугу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Обновить услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные услуги",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Service"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Service"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет услугу.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Удалить услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}/assign-master": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Привязывает услугу к мастеру.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Назначить услугу мастеру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Пользователь-мастер",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AssignServiceToMasterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}/assign-master/{profileId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отвязывает услугу от профиля мастера.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Убрать услугу у мастера",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID профиля мастера",
                        "name": "profileId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AssignServiceToMasterRequest": {
            "type": "object",
            "required": [
                "target_user_id"
            ],
            "properties": {
                "target_user_id": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "bad_request"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
...
        "models.EmployeeProfile": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expected_salary": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "salons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Salon"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "specialization": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "work_schedule": {
                    "description": "{\"mon\":\"9-18\",\"tue\":\"9-18\",...}",
                    "type": "string"
                }
            }
        },
        "models.Material": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "description": "мл, шт, гр",
                    "type": "string"
                }
            }
        },
...
        "models.Salon": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "base_hourly_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EmployeeProfile"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "description": "PostGIS, опционально",
                    "type": "string"
                },
                "max_staff": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "OPEN/CLOSED",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "working_hours": {
                    "type": "string"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EmployeeProfile"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServiceMaterial"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ServiceMaterial": {
            "type": "object",
            "properties": {
                "material": {
                    "$ref": "#/definitions/models.Material"
                },
                "material_id": {
                    "type": "integer"
                },
                "quantity_per_use": {
                    "type": "number"
                },
                "service": {
                    "$ref": "#/definitions/models.Service"
                },
                "service_id": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "employee_profile": {
                    "$ref": "#/definitions/models.EmployeeProfile"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.UserRole"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "CLIENT",
                "BASIC_MASTER",
                "ADVANCED_MASTER",
                "HR_SPECIALIST",
                "ACCOUNTANT",
                "NETWORK_MANAGER",
                "ADMINISTRATOR"
            ],
            "x-enum-varnames": [
                "RoleClient",
                "RoleBasicMaster",
                "RoleAdvancedMaster",
                "RoleHR",
                "RoleAccountant",
                "RoleNetworkManager",
                "RoleAdmin"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}